move MaximiseToggle() to its own file
